+++ frontend/styles/campaigns/detailStyles.ts
@@ -1,5 +1,5 @@
 import { SxProps } from '@mui/system';
-import { theme } from '../../../utils/theme';
+import { theme } from '../../utils/theme';
 
 export const donationBreakdownItemContainer: SxProps = {
   alignItems: 'center',
@@ -15,10 +15,6 @@ export const charityLogoSx: SxProps = {
   maxHeight: '48px',
   minHeight: '48px',
   height: '48px',
-};
-
-export const leftBarSx: SxProps = {
-  backgroundColor: theme.palette.primaryTranslucent.main,
 };
 
 export const rightBarSx: SxProps = {
