+++ frontend/styles/components/redeem/ReceiptStyles.ts
