--- frontend/styles/redeem/indexStyles.ts
@@ -1,10 +1,0 @@
-import { SxProps } from '@mui/material';
-
-export const itemListSx: SxProps = {
-  height: '450px',
-  cursor: 'pointer',
-};
-
-export const itemSx: SxProps = {
-  height: '500px',
-};
